% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SBGridCapsuleLogs-package.R
\docType{package}
\name{SBGridCapsuleLogs-package}
\alias{SBGridCapsuleLogs}
\alias{SBGridCapsuleLogs-package}
\title{SBGridCapsuleLogs: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Thomas Nicholson \email{nicholson.tom.n@gmail.com}

}
\keyword{internal}
